Simplify codegen of binary ops involving closures, such as Ci = val * closure()
In OSL's ancient past, copying closures was expensive, so where possible
we tried to generate code that looked like

    result = leftoperand()
    result = result OP rightoperand()

rather than the more straightforward

    lefttmp = leftoperand()
    righttmp = rightoperand()
    result = lefttmp OP righttmp

But C (and therefore OSL) semantics dictate that the operands be fully
evaluated before the result is altered, and it turns out that we missed
detecting a couple cases where this could be violated, in particular if
the leftoperand() itself assigned to result or if rightoperand() used
the value of result (we got that right for +, but not * or /), or if the
result is a global (like Ci) in which case evaluation of leftoperand
could involve a parameter that triggers execution of an earlier layer
which itself uses or sets the same global.

But since we've long since made a closure copy merely be a pointer copy,
this is all pointless!  So greatly simplify the code while switching to
the opcode sequence that is guaranteed to be correct.